# services/backend-api/openapi.yaml
openapi: 3.1.0
info:
  title: GoMaths Backend API
  description: |
    Core REST API for GoMaths. This file is the source of truth — `@gomaths/api-client`
    is generated from it via `openapi-typescript`.

    Versioning policy: breaking changes bump the major version path (e.g. `/api/v2/...`).
    Phase 1 ships as v1, served at the `/api` prefix.
  version: 0.1.0
  contact:
    name: GoMaths Engineering
servers:
  - url: https://api.gomaths.co.za
    description: Production (af-south-1)
  - url: https://api.staging.gomaths.co.za
    description: Staging
  - url: http://localhost:4000
    description: Local
...
security:
  - bearerAuth: []
paths:
...
  /api/tutor/messages:
    post:
      tags: [tutor]
      summary: Send a message to the AI tutor
      description: |
        Proxies to `ai-services/tutor`. The response is SymPy-validated before
        being returned to the caller. Rate-limited to 20 requests / minute
        per student.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TutorMessageRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TutorMessageResponse" }
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/tutor/messages/stream:
    post:
      tags: [tutor]
      summary: Stream a tutor reply over Server-Sent Events
      description: |
        Returns `text/event-stream`. Event types:
          - `meta`   `{ conversationId }`            (always first)
          - `delta`  `{ text }`                      (zero or more)
          - `claim`  `{ raw, stem, answer, ok }`     (zero or more — one
                                                    per equation as soon
                                                    as the SymPy validator
                                                    has scored it)
          - `done`   `{ reply, validated, provider, model, extracted_claims,
                       verified_claims, input_tokens, output_tokens,
                       cached_tokens }`
          - `error`  `{ message }`
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TutorMessageRequest" }
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream: {}
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /api/tutor/conversations/{id}:
    get:
      tags: [tutor]
      summary: Load a conversation by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Conversation" }
        "404":
          $ref: "#/components/responses/NotFound"
  /api/curriculum/check:
    post:
      tags: [curriculum]
      summary: Check a student's answer to a curriculum question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [questionId, answer]
              properties:
                questionId: { type: string }
                answer: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [questionId, correct, validated, expected]
                properties:
                  questionId: { type: string }
                  correct: { type: boolean }
                  validated:
                    type: boolean
                    description: |
                      True iff the SymPy validator was reachable AND the
                      answer was symbolically evaluated. False means the
                      check fell back to literal string equality.
                  expected: { type: string }
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
...
    TutorMessageRequest:
      type: object
      required: [message]
      properties:
        message: { type: string }
        topicId: { type: string, nullable: true }
        conversationId:
          type: string
          nullable: true
          description: Omit to start a new conversation.
    TutorMessageResponse:
      type: object
      required: [conversationId, reply, validated]
      properties:
        conversationId: { type: string }
        reply: { type: string }
        validated:
          type: boolean
          description: |
            True iff every numeric/algebraic claim in `reply` passed SymPy
            validation. False means the response is best-effort prose only.
...
    Conversation:
      type: object
      required: [id, studentId, createdAt, updatedAt, turns]
      properties:
        id: { type: string }
        studentId: { type: string }
        topicId:
          type: string
          nullable: true
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        turns:
          type: array
          items: { $ref: "#/components/schemas/ConversationTurn" }
    ConversationTurn:
      type: object
      required: [role, text, occurredAt]
      properties:
        role:
          type: string
          enum: [user, maya]
        text: { type: string }
        occurredAt: { type: string, format: date-time }
        validated:
          type: boolean
          nullable: true
          description: |
            True iff Maya's reply was fully SymPy-verified. Only meaningful
            when role=maya.